Refactor the cache_promote policies
