+++ plugins/cache_promote/Makefile.inc
@@ -17,4 +17,7 @@
 pkglib_LTLIBRARIES += cache_promote/cache_promote.la
 
 cache_promote_cache_promote_la_SOURCES = \
-  cache_promote/cache_promote.cc
+  cache_promote/cache_promote.cc \
+  cache_promote/config.cc \
+  cache_promote/policy.cc \
+  cache_promote/lru_policy.cc
